ml/requirements.txt
# OCR and document processing
pytesseract==0.3.10
Pillow==10.2.0
pdfplumber==0.10.4
camelot-py[cv]==0.11.0
PyMuPDF==1.23.8

# ML frameworks
torch==2.1.2
transformers==4.36.2
datasets==2.16.1

# Data processing
numpy==1.26.3
pandas==2.1.4
opencv-python==4.9.0.80

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Fuzzy matching
rapidfuzz==3.6.1

# ML training
scikit-learn==1.4.0
lightgbm==4.2.0
joblib==1.3.2
faker==22.0.0

# LLM integration
openai==1.12.0

# Testing
pytest==7.4.4

